requirements.txt
aioboto3==12.0.0
aiobotocore==2.7.0
aiohttp==3.8.6
aioitertools==0.11.0
aiomysql==0.2.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncmy==0.2.8
attrs==23.1.0
boto3==1.28.64
botocore==1.31.64
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
fastapi==0.104.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
motor==3.3.1
multidict==6.0.4
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pymongo==4.5.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
sqlparser==0.0.9
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
wrapt==1.15.0
yarl==1.9.2
